Added a test application for testing how different diff tools work in different scenarios and added beyond compare
