+++ src/Shouldly.Shared/Configuration/ITestMethodFinder.cs
@@ -1,0 +1,11 @@
+#if !PORTABLE
+using System.Diagnostics;
+
+namespace Shouldly.Configuration
+{
+    public interface ITestMethodFinder
+    {
+        TestMethodInfo GetTestMethodInfo(StackTrace stackTrace, int startAt = 0);
+    }
+}
+#endif
