+++ src/TestDiffTools/Main.approved.txt
@@ -1,0 +1,1 @@
+Foo
